requirements.txt
streamlit
opencv-python
opencv-contrib-python
openpyxl
pandas
numpy
Pillow
requests
scikit-learn
